Final Homework Upload
I understand that the energy and magnetization plots do not look exactly like plots provided in the Rajesh Singh link. Some corrections had be made to the Ising model code (Lecture 18 code) to produce the plots shown in the uploaded file (otherwise the plots were just noise). I am not sure but there may be some remaining issues with the Lecture 18 code that were not fixed.